Restructure client & remove embedded controllers
